OLD/requirements.txt
flask
boto3==1.11.9
botocore==1.14.9
docutils==0.15.2
jmespath==0.9.4
python-dateutil==2.8.1
python-dotenv==0.10.5
s3transfer==0.3.2
six==1.14.0
urllib3==1.25.8
